requirements.txt
# Core Framework
crewai>=0.28.0
crewai-tools>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.20

# LLM
ollama>=0.1.6

# Web Scraping & Automation
selenium>=4.16.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
browser-use>=0.1.0
requests>=2.31.0
httpx>=0.25.0
lxml>=4.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0

# API Clients
ShopifyAPI>=12.7.0
google-trends-api>=0.0.1
pytrends>=4.9.0
tiktok-api>=0.1.0

# Image Processing
Pillow>=10.1.0
opencv-python>=4.8.0

# Database
sqlite-utils>=3.35

# Validation & Models
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
tenacity>=8.2.0
fake-useragent>=1.4.0

# Analysis
textblob>=0.17.0
vaderSentiment>=3.3.2
wordcloud>=1.9.0

# Export
reportlab>=4.0.0
jinja2>=3.1.0
markdown>=3.5.0
